generate all keys for each dataset
In order to give users a better idea of how they
should create one of these config files if they
do it by hand, I changed the config file
generation to add all keys for each dataset even
if they aren't specified on the command line,
similar to how the rest of the config file
generation works for plain options.

Also removed global get_object_key function as
it is now part of the GradientObjectListOption
class.