Add basic session management
This introduces a pretty lightweight overlay to enable session
management within the Python client. The idea here is that in many
non-trivial workflows, users will want to switch among different
environments, or projects, or access levels easily and intuitively.
E.g., being able to run code in a nonprod environment, validate it, and
then elevate it into a production context.

For a mature data science team, these different environments will likely
be separated and have different sets of credentials to enable
interaction with them. (An analogous real-world example here would be
the AWS boto3 Python library, where a user can instantiate multiple
`boto3.Session()` instances, each with separate credentials and IAM
profiles.)

There are a few ways we can enable this, but given some of the drawbacks
of the previous iteration of this package I wanted to try and avoid a
super class-heavy approach where we have client classes and wonky
inheritence structures. So puttered around with a few alternatives and
ended up settling on what we have here: a `Session` class and a function
decorator that can wrap our http adapter. It needs a little bit of
finessing still since I'm basically just always pulling the default
session but it should be relatively straightforward to fix up.

The big draw of this approach for me is that all of the client methods
we implement can be totally agnostic to user credentials, which
moderately simplifies the interaction pattern versus the GQL methods
from V1 of the package, and pretty considerably simplifies things from
the class-heavy clients that we're migrating away from.